0021360: [AppVerifier] Crash when making muvee, using a mp3 or mpg file
The video frame grabber filter that's created as part of the GraphForFile() function call
is destroyed if the file only contains audio. However, the open() function call (line 289)
was referring to m_fg which contained a dangling pointer. 

This commit makes the irrelevant pointers NULL - i.e. m_ag will be NULL for
a video-only filter graph and m_fg will be NULL for an audio-only filter graph.
This makes it easier to place assertions. The check at line 289 now uses the
NULL value and accesses the right pointer to determine whether to pause/start 
the stream.

With this commit, the first access violation (under AppVerifier) for  mp3 and such files 
is fixed.