0021789: [AppVerifier] Potential problems with MESCore.dll
Modified MESCore.dll and the MESCore class in the utils project
to perform separate initialization and termination of the library
using the MESInit() and MESShutdown() API calls.

The effects that are included in mescore.dll initialize themselves
using a static object for each one of them. I've moved these
static objects into an explicit initialization routine FX::RegisterCoreEffects()
that is called by the MESImpl class upon MESInit().