RESOLVED: #0021007: Glitch in dissolve between first and second picture in the presence of a title - in Classic Vanilla style
The issue here was that there were two cross-fades were occurring at the same time. 

The title to 1st segment transition started at 3.5s and ended at 4.5s.
The 1st segment and the 2nd segment transition started at 4.4s and ended at 5.3s.

Because the title segment and the 1st muvee segment are completed different shows, what happened was that there were two instances of segment1 during the two cross-fades. The first one was still fading-in while the second segment1 was fading-out. When you added up the alphas of the two segment1, segment1 briefly had an alpha of 1.0.


This issue has been resolved by not using the "CrossFade" transtion but instead using a custom cross-fade from within muse. The custom cross-fade has a small z-offset between the different layers.