0021449: Preview unable to start on some HP machines
The commit 6238 (~57977) for this issue was merged to Reveal_BR1_branch to make the opengl code path up to the start of preview identical between the trunk and the branch. There is an access violation that occurs with AppVerifier turned on on BR1, but not on the trunk. This commit doesn't seem to resolve the access violation, but without it, what's on BR1 as of this commit won't work on said HP machines.