0021775: Saving and loading mES timelines
The names of the timeline components are now chosen
based on the type of the component rather than
just s9 s123 s241 etc. So you'll now see layers15,
effect-track32 and such. 

Additionally, effect terms are named after the actual 
name of the effect, so you'll also see Alpha23, Translate42, 
etc. This is slightly better than the previous uniform naming 
scheme.

The labels for portions of the timeline are now all written
to the end of the file. So the end of the file is the best
place to start navigating the show graph.