Resolved: #0021933: [Spring09] (hp# 80616) PEGATRON: ASF format video can not be added after adding music first in muvee Reveal.
ASF extension is added for both video and audio. So when main app count for new added music, it counts it as both audio and video. As asf is most likely to be video, so we will count it as video (by using else if, instead of if for music to avoid double counting).