0020700: Handling muvee duration better
(WORK-IN-PROGRESS)

A modified constructor algorithm to address incorrect "not enough media" 
issues is in the works and enabled using c:\muveeDebug\UseInflator.txt.
The reason for the name is that it works by inflating a picture-only 
production with video until you hit the target duration.

The core case that works for cases that didn't work before is when there
are lots of pictures and video clips as shot from a DSC. 

Limitations of the UseInflator.txt mode (basically a lot at the moment)

* Input shots have to be longer than style generated segment durations in order to be included. (fiddly but can)
* Highlights are not accounted for (in progress)
* minimum transition duration setting in style is not respected. (easy)
* min seg duration for transitions is not respected. (easy)
* account for mismatch between transition duration TC and transition selector - i.e. when the TC says "trans" but the selector says "cut". (easy)
* picture stretch/shrink/drop to fit to duration. (can do by just iterating the algo)