Resolved: #0021943: [Spring09] (hp# 80362) There is no Muvee short cut on desktop after closing Muvee Reveal by clicking "Add Shortcut on desktop"
[DETAILS]:
Coordinator::CreateDesktopShortcut has a code path for the condition where the user languages were less than 1. This condition was expecting a string table called 'ShortcutNames.txt'. Modified this to use the strings.xml string table instead, and also to install the shortcut the vista way, when the OS is vista, and the XP way when its not (no more language counting, just os checking).