0021930: [Cube Twist] Jerky playback followed by crash when video background is used.
changes merge from Reveal_BR1_branch rev: 6954

Some time seeking to start of the file failed and we try to set the the seek point to current postion of the Graph, apparently seeking shouldn;t be failed if the seeking point is within the druation of the file. Seeking is improved by setting the Graph to pause mode then seek to the point and then run the Graph.