0022046: Pan/Zoom for Portrait Pics No longer requires that the selected portion of the picture entirely fills the output frame.  If there's a sizable mismatch between the picture's aspect ratio and the output frame's aspect ratio, we allow "black bars" to appear.  Black bars are shown if not showing black bars would mean that more than some fraction of the picture is "lost", i.e. chopped off the top/bottom (if the picture is tall compared to the output frame) or sides (if the picture is very wide).
There are separate "fracLost" thresholds for horizontal and vertical, so one could say allow bars on top & bottom if over 10% of the pic would be lost from the left & right sides (combined), but allow on the sides if over 20% of the pic would be lost top/bottom.  At the moment, both thresholds are set to 0.2 (20%).  For experimentation, they can be overridden with C:\muveeDebug\FracLostThresh.txt. See bugnotes for details.