0021663: [TC39-01-204] App stops responding after returning from hibernation. 0021733: [TC39-01-212] After recovery from sleep mode, application hanged 0021736: [TC39-01-215] After recovery from sleep mode, application hanged 0021737: [TC39-01-216] After recovery from Hibernation mode, application will hanged 0021662: [TC39-01-203] App stops responding after returning from sleep.
After recovering from sleep, (priot to this commit) recording any voice over in the VO dialog and then pressing stop will result in the loss of the recorded bit - i.e. no voice over is recorded. This fix re-enables voice over recording by resetting a flag that was causing the voice over clip to be discarded.