#22150: Prevent trimming of whitespace for XML element.
Minor tweak to the code in inicrawl to l-strip and r-strip 
style search paths of white space so that style enumeration
behaviour is preserved. We need to do this now since the default
setting has been changed to "preserve white space" and therefore
newlines also get included in the paths returned by AsText().