Resolved: #0020810: [DEU] When importing from camera, if the folder consist of duplicate pictures, error msgbox that pop up is not localised
Added message box title, and set language when engine is loaded.
When testing, please rename two file to be same name. Please do not just copy file to a sub folder, addon is intelligent enough to detect that there are same file and it will not treat as duplicate file. Duplicate files mean file with same name but different file time, etc.