0022180: Application crashes after starting titles preview once (in personalize), cancelling out and re-entering.
When the preview is shutdown, an exception might be thrown
within the renderer funciton, say, when it is waiting
for a video frame from the video decoder. The thrown
exception will be FInterruptedEx. Upon encountering
the exception, the preview renderer should do the proper
timeline cleanup operations - in particular show->Leave().

Since this exception was not being handled in the Receive()
function of the preview renderer, it was being propagated
all the way to FThread, causing the show->Leave() to
not be called in the correct opengl context.

This fix catches the FInterruptedEx in the preview
renderer's Receive function so that proper cleanup
actions are performed before exiting the opengl
context.

(merged from Reveal_BR1_branch -c 7109 to trunk )