Resolved: #0021930: [Cube Twist] Jerky playback followed by crash when video background is used.
The AccelOp operator doesn't need to hold on to a reference to
the last received video frame. Because it did, the Touch() call
at the start of every render frame resulted in a delayed VideoFile
object to be destroyed before AccelOp's Leave() got to be called.
That meant the VideoFile was destroyed before one of the frames
from it got to be released, resulting in a crash. 

Not holding on to a reference to the VideoFile's pixels is enough
to fix this. That is safe since if the VideoFile can be safely destroyed,
you're guaranteed that frames from the video file are not needed,
according to the Touch() function specification.