Resolved #0022247: [Spring09] (hp# 82147) 21115_muveeReveal_B1925P_01_VIS_ALL+WW-01:  Muvee Preview Black screens, then will not drop from Task manager
In this case, the user presses pause/play rapidly, causing the timeline 
to perform audio seeking to nearby locations rapidly, causing a hang in 
mediaSeeking->SetPositions(), which is generally not supposed to hang for
any reason. So we prevent these SetPositions() calls by first checking the
cache for the relevant samples and issuing seek requests for only those
parts of the stream that are not in the cache.