Resolved: #0022193: [Spring09] Burning with 3 MPEG2 files + 3 WMAs with Pump It Up will crash
newfx_layers function in constructor was returning MUSE_NIL
for very short effect instances (7ms in this case).
Earlier, newfx_effect was modified so that it will never be
MUSE_NIL, but that contract was not propagated to newfx_layers
at that time. Making newfx_layers behave the same way as
newfx_effect for ultra-short effects resolves this problem.