Resolved #0022253: memory leak from DirectShowVideo.dll wzdebug.cpp Resolved #0022254: memory leak in DirectShowVideo.dll amfilter.cpp CBasePin Resolved #0022255: memory leak in DirectShowVideo.cpp, GraphForFile( ) Resolved #0022256: Memory leak in DirectShowVideo.dll renbase.cpp GetPin
While construction of Video / Audio decoding Graph , both audio and video grabber filter are initialized for each stream and later if audio stream then video grabber is removed but reference count of smart pointer was not 0 at the end of decoding graph that why even if we remove audio or video grabber filter it not completely  removed because of its reference count.

We have modified the code to make sure reference count is always 0 at the end of the decoding graph construction.