0022065: [Spring09] AppVerfier [Intel graphics card] : "UserDefined breakpoint hit" after cycling the title / credits tab in personalize 0022054: [Spring09] AppVerifier [Intel graphic card] : "UserDefined breakpoint hit" upon making a second muvee 0022053: [Spring09] AppVerifier [Intel graphics card] : "UserDefined breakpoint hit" upon making a muvee, and the clicking on 'stop'
Intel drivers crash (under AppVerifier) at glClear on soem occasions. 
I found that if the glDeleteTexture call that is made in STextOut is 
delayed until the next render cycle - by moving it into the cooperative
".OpenGL" thread - the AppVerifier access violation crash disappears.

As far as I understand, doing it here or delaying it are equivalent
in terms of opengl specification, so I don't know why one works with
Intel drivers and the other doesn't. Both calls are made in the same
thread, under the same render context, associated with the same window.

So, I'm now delaying the release of the text effect's texture as a workaround
for the AppVerifier crashes we see on Intel OpenGL graphics drivers.