0019759: muFAT crashes and assert failures - The Collection 0021775: Saving and loading mES timelines
Saving the timeline generated by PumpItUp style during one 
of the muFAT runs was causing muSE to generate new symbols
beyond what is set as the maximum capacity. The fix to
the "constructor_allstyles" and such cases crashing on all
machines due to timeline saving is to allow muSE to grow the
storage space for defined symbols on demand so that it never
hits the limit and doesn't know what to do.

Note: I took the chance to remove a related but dead/unused 
static function ensure_stack() in muse.c.