Resolved #022330: mR unable crash on ws-sohrab.sg.muvee.net machine
Problem is resolved by checking the number of output pin in source filter , if there are more than one output pin we assume that output pin will be one video pin and one audio pin in this case we will get the pin based on pin type and render it with proper rendering filter either audio rendering filter or video rendering filter.

If there is only one output pin in source filter then we just render it directly.