Resolved #0022332: [Spring09] (hp# 82665) Pegatron: Muvee Reveal stops working after cancelling the saving process
Made AccelOp not use itself for remembering last used texture.

---8<--- copy pasted from source code ----8<----
It is possible that when video is played
in slow motion, AccelOp's Eval() might receive
the same pixel image for multiple output frames.
In such case, it should not spend time re-uploading
the pixel image to an opengl texture and should
simply reuse the texture it used for the
previous pixel image. The TexCache object is
intended to serve this purpose - to map the AccelOp
Show to a saved texture that it can refer to.

The sole TexCache instance for a given render
environment is stored in the OpenGL thread that
the environment provides, in the thread's dictionary,
associated with the "TexCache" key. When the OpenGL
thread gets destroyed in the proper GL context,
all the texture stored in its dictionary will
also be released properly before deleting the context.

This implementation gets rid of all memory in the
Show timeline of having been rendered in some 
environment, therefore allowing the same timeline
to be rendered in multiple environments with
correct resource management.
---8<--------------------------------------------------