Resolved #0022323: Stuck in "Buffering" indefinitely when previewing series of muvees with DV input
AudioGrabber.cpp / FrameGrabber.cpp
Sample Rendering time set to NULL to render sample as fast as possible

DirectShowVideo.cpp
Release all the pending samples before Stop and relasing the DirectShow GraphForFile

ShowTime_Renderer_FullScreen.cpp

Make the Die() function operate directly within the preview render loop.
It now sends a message to the preview thread to die, which gets
received within the preview loop. It then waits for the renderer's
Receive() function to return before proceeding to declare the
renderer as "dead".

Modify the idle call within the preview renderer's loop indicate using
a boolean value that the idle calls are being made within the
loop. This is so that the Die() function can distinguish whether
it is running within the loop or whether the loop has finished
and the renderer's Receive() function has returned.

MVCore.cpp

Release the previous muvee's timeline *before* constructing a muvee.

TLEngine.cpp

Release the previous render environment *before* construction a muvee.

FMutex.cpp

FWaitForEvent() function needed to be modified to
accommodate unexpected behaviour of the MsgWaitForMultipleObjects
function. What was happening is that MsgWaitForMultipleObjects was
returning with a value that indicated the presence of a message in the
message queue, but a subsequent PeekMessage() returns FALSE,
indicating that there is no message in the queue. This caused the
Wait() operation in the renderer's Die() function to return prematurely,
causing the construction to start before the previous render environment
was completely destroyed.
	
Therefore, FWaitForEvent() needed to be modified to accommodate the
weird PeekMessage() behaviour. I'm sure there is an explanation for it,
but it sure doesn't seem straight forward. The modified

FWaitForEvent() will continue to wait in a loop until WM_QUIT is posted to the waiting thread
using PostMessage().  I think this is the expected behaviour for
FWaitForEvent() in all cases used within mES, so we need this fix across the board.