0019759: muFAT crashes and assert failures - The Collection
baseline-preview.run was crashing on the second preview.
The crash was due to a new test of a situation that doesn't
occur in Anchorvale, although it is desirable to support that
from a design perspective - a given ShowTime timeline could
not be rendered using two different render environments.

This suite of changes makes a timeline renderable in as many
render environments as deemed necessary within a single session
without requiring a reconstruction.