#0022009: S00532_BumperPix  	 #0022010: S00533_FancyFrames  	 #0022011: S00534_Mystical
Because the 3 styles of photoGenie had been given new names and descriptions, I deleted all the previous localizations for Style name and style descriptions.

It turns out we can keep them. In this way, we can say that we have a fully localized pack.

Only downside is S00532_BumperPix is known as Grunge in english but BumperPix in all other languages. It's rather harmless as practically noone will know that.