Resolved #0022411: Cannot save muvee with a quicktime supported file added twice in the media panel.
changes merge from trunk rev 7399 

Failing of QuickTime is not related to FindFrame time out.

QuikcTime was failing second time for same file because file was being opended in Read and Write mode so opening same file twice in write mode cause quikctime to failed . 

Now QuikcTime file will only be opened in read mode because there is no need for file to be opened in write mode when we are decoding the file.