0020700: Handling muvee duration better
(WORK-IN-PROGRESS)
(UseInflator.txt)

There is a limit to the duration of the muvee
that can be made given input video without any
attempt to stretch or shrink the timing. If the user
specifies a manual duration (or music duration)
that is longer than this, the constructor (in
previous version) will try to accommodate that 
by stretching the segment timing. Doing so 
increases segment duration and decreases the proportion
of time spent in transitions and can therefore
achieve some extra duration. However, this is not
desirable because you don't get a natural looking 
production as well as duration is not satisfied.

Therefore I put in an additional case that if
the constructor is not able to accommodate the
higher manual duration by stretching, then it 
simply makes a muvee at the natural pacing and
accepts the fact that the muvee duration is shorter
than the specified duration - only one evil instead
of two.