Resolved: 0022512: [SpringRefresh09] Application is buffering non-stop after switch styles and make muvee for few times
AudioFile was getting stuck in SubmitFragmentRequests() in an infinite
loop if RequestFragment didn't do anything. So modified the function
so that if RequestFragment didn't do anything because the stream is
in MV_VIDEOSTREAM_LIMBO state, SubmitFragmentRequests() immediately returns.

This fix has been checked on Yanxi's HP machine. Hence confident of
resolution.

(merged v7604 from trunk to Reveal_BR2_branch).