0020700: Handling muvee duration better
(WORK-IN-PROGRESS)
(UseInflator.txt)

For video files containing short segments, the inflating 
constructor wasn't supporting shots that were shorter than 
required segment duration and in those cases it might
do a slow motion stretch.

For the moment, I'm preventing that from happening. The downside
is that we might now have shot boundaries straddled. This is flagged
by an assert failure. In normal cases, shot boundaries are unlikely
to be crossed since there is a bias against it.