Resolved: #0022571: [SpringRefresh09] Cannot burn into DVD with chinese character inside (unicode character)
This is because that the special writer used to fix the 1GB burning issue (which will split the file into smaller parts so that burning can succeed) does not support unicode. So for burning, we will pass none unicode file name to writer, and then rename it back after saving completed.