Resolved #0021332: Clicking on saving dialog when saving reaches 100% causes the application to stop responding.
When Saving to any format progress is always in between (0 to 1 ) when progress reaches  > 0.995 total % calculator round up to 100% and at that point saving is not completed yet.

Now when progress reaches > 0.995  percentage will always remain 99% untill we actually receive signal from SDKRuntime that saving is completed by returning progress 1 then we set the percentage to 100%. 

MP4 and 3gp files are more effected because QuikcTime always generate the mov files and when we are about to destory the Graph conversion is perfomred from mob to mp4/3gp.