0020700: Handling muvee duration better
(WORK-IN-PROGRESS)
(UseInflator.txt)

When shrinking pictures to fit to a duration 
(not playing any video), the muvee duration
usually ends up a bit different from the target
duration. In this change, I'm trying a different
picture shrinking algorithm - keep dividing
the longest segment into two until you can fit
all pictures and then adjust the last segment
to match the target duration exactly.

I'm not sure I like this approach 'cos the
pacing of the muvee is not uniform, but the
previous algorithm was causing a larger 
deviation than necessary and I think this one
does better.