Resolved #0022458: [SpringRefresh09] Saving to MP4 will hang at 99%
Wheen saving to mp4 ro 3gp format there is post process of format conversion after saving is complete this is done at the point when we stop the Graph, and stop graph function is called inside the Get Progress which is actually called from UI this will cause UI to freeze untill stop graph is successfull.

Now get progress is not blocking call stop graph will be called from the timer which actually update the progress bar , releasing of saving graph and mes objects will still be done from Get Progress bar function.

This will help to reduce the UI freeze when there is post processing of format is going on after saving is complete UI progress will never be set to 100% untill Stop graph is called.