0013043: muSE-X related changes
Effect expressions could not be used in the value part of an
(input N show) expressions because newfx_effect wasn't saving
and restoring its context when evaluating its body. Originally,
the input expression was only designed to allow completely
defined Show objects to be used (such as those created by
mes:video, mes:image, etc.) Its kind of nice to be able to
use effect expressions as well and so I added context save/restore
to allow that.