Resolved: #0019571: [TC39-01-225] During burning, application hang when resume from Hibernate
We did not do anything to prevent system from going to sleep during saving, but it seems that during saving, system is so busy that it can not go to sleep. That why when Nicholas cancels saving, system goes to sleep immediately.

Change the behavior to prevent XP system to go to sleep during saving.