Resolved #0022583: [SpringRefresh09] (hp# 83624) MuveeReveal (21447): Previewing a WMV video and clicking on “ok” button will cause audio of the vi
When closing the preview video window Unique id for the active graph was being reported wrong because before we close the preview video window app store video info for all the input video file and the counter which represent the input video file is move to end that is why when we actually close the graph that counter represent the wrong FilterGraph.

We are closing the video graph first then stroing the information of all input  video files.

Changes made by yeecheng..