Resolved: #0022604: Slomo segments in KIA production with Ultra Plain
One of the playaback speed computation fixes made for the normal construction
cases had not propagated to the use-it-all mode. Did it and it solved it.
There was also a small accumulation error that crept into the computation
of the time warp map that I also fixed to make the LoopIntegral behave
predictably. Previously, for the exact case of 300seconds of music, it
showed a drift error of 1/1500. The error will actually not be noticeable
due to the use of the LockedPlaybackSpeed() function which will lock
speeds close to 1.0 to 1.0 in order to avoid sync sound drift.