0013043: muSE-X related changes
1. Made title/credits construction cancelable upon muSE errors.

2. Expressions such as (segment-durations ...) are not relevant to
title/credits construction, so I put in special cases to not evaluate
such expressions which have an undefined symbol as the head of the expression.
This lets me strengthen the error reporting in generic muSE by checking
for undefined symbols used in function call position.

3. Moved the setup of default exception handler to MuseUtils so it can
be shared between title/credits construction and body construction.

4. Made StyleParamInfo evaluate all expressions before the (style-parameters ..)
expression, ignoring unprocessable expressions. This lets us define functions
used to generate the style-parameters expressions if we need to. Additionally,
it lets you use tab-syntax for style-parameters as well.

5. Strengthened the error message presented when an undefined symbol
is used in function call position. Prior to this suite of changes, I
couldn't do it this way because the style parameter parsing, title/credits
parsing all had to deal with undefined functions (ex: segment-durations)
that are relevant only to the muvee body. Now we're clean since those
cases ignore such expressions.