0013043: muSE-X related changes
Many of our current styles apply undefined functions
on argument lists during the parsing of the scm file
for info about style parameters and title and credits
defaults. Therefore it is unfeasible to leave in the
strong check for undefined function applications 
in muSE because they will break currently installed
styles when the user updates muvee Reveal.

This commit performs the "undefined symbol" check
only if the file c:\muveeDebug\muSE_BlockUndefinedSymbols.txt
is present. Otherwise it silently ignores the application by
turning the application into a list, as it had done
in earlier vesions.