0022560: Disable AMM on low end GPUs.
Behavior changes.

Previously, a global decision is concluded in the initial performance test and it's stored in the OpenGL thread. As a result, all textures associated with the OpenGL thread will be created with AMM support if and only if the earlier global decision is "good".	 

In this commit, the thing remaining the same is that the same performance test is carried out before the first FImageTexture object is created. Then, instead of a global decision ("good" or "bad"), the average cost of executing glTexSubImage2D() on a 512X512 texture is stored in a global variable. Now whenever a new FImageTexture objected needs created, an extra parameter "timeLimit" is added which specifies the maximum duration allowed for the creation. AMM would be enabled on a texture if the "timeLimit" is greater than the glTexSubImage2D cost.

The reason of this is that we only want to disable AMM when we have to, because enabling AMM reduces artifacts. Also Gerry suggests that his slow XP computer runs photo-only muvee well, but not with video inputs. Therefore, we should still try to enable AMM for texturizing photos as long as it not too slow (For texturizing non-video frames, we give generous timeLimit=200ms).