0013043: muSE-X related changes
A muSE module can itself be used as a function
and it behaves like "do" except that the body part
is interpreted relative to the symbols exposed by the
module. In this commit, I generalize this notion and
let the module define how to handle this condition
by defining a "main" function within it (which may 
or may not be exported by the module). This serves
as a very easy route to defining small custom languages
within  muSE itself.