0013043: muSE-X related changes
{tab-syntax}

Made the empty group turn the previous term
into a singleton list containing the term.
This gives us function-call-like notation
for calls without arguments and lets us use
"a()" instead of "a : []". Since "((((a))))"
means the same thing as "a", the meaning
of "()" is open to interpretation and we
are justified in using it this way.