0017522: Eliminate build warnings
The problem with the MuseUtils.cpp warning message
was that MuseUtils.cpp had optimizations turned on
in Release build whereas every other module in 
utils has optimizations turned off in release build.
Hence the precompiled header was incompatible with
the code generated for MuseUtils.cpp.

So I've changed MuseUtils.cpp's optimization settings
to off as well to be consistent with the rest of the
files. Is this "foolish consistency", I don't know,
but will do for the moment.