StyleLocChecker
Just drag and drop style folders to StyleLocChecker.exe and let it check through the encoding validity of localization strings. It generates an HTML summary at the end of the process. Invalid encodings will be highlighted red.

Works with aP6 and mR style folder structures, as well as both LCID_XX and yy-ZZ locale codes.