0020700: Handling muvee duration better
{WORK-IN-PROGRESS}
{UseInflator.txt}

The GetEndShotIx() function as written could get into 
an infinite loop under certain conditions. Rewrote
the function recursively so that at least the compiler
will tell me when I'm writing an infinte loop, of course,
fixing the infinite loop in the process.

Also reduced the low quality descriptor threshold from 
0.1 to 0.04. Karen's material on the "poi", for example,
still has some reasonable motion in it and is still usable,
but the quality descriptor hovers around 0.1 for those
video clips. Considering that much of the material is
going to come from DSCs and HD camcorders, it makes 
less sense to have such a high low quality elimination
threshold. In fact, I'm even tempted to set the
threshold to 0.0. But anyway, for now, lowering it
is still ok and it satisfies the "blue screen" cases
present in DV camcorders.