0013043: muSE-X related changes
Changed the default value of transition duration
transfer curve from 0.0 to 1.0 as per Teng Chwan's
request. The rationale is that if a style author
only defines "muvee-transition", he expects to
see the transition in the muvee. The default for
preferred-transition-duration is in alignment with
that principle but the default for transition-duration-tc
wasn't. So brought that into alignment :)