0019759: muFAT crashes and assert failures - The Collection
The saving crashes happening in muFAT for the past two runs
on the trunk seem to be happening due to a texture being
released outside of an opengl context. The interesting thing
here is that this problem doesn't happen if you don't have
the MESHardDebug.txt flag (which muFAT sets)!! So I suspect
it has to do with FSCOPE related changes.

However, on probing the save renderer, I found that it could
have used ReleaseResourcePools() call in a couple of spots.
When I added that, the saving crash disappears even with
MESHardDebug.txt on. So I'm committing the "safer release"
change to the save renderer.

The change in GLResourceManager accompanying the save renderer
change is to allow the Destroy() function to be called multiple
times without problem (which I usually allow).

(merged change in 7951 from trunk to Reveal_BR2_branch)