0019759: muFAT crashes and assert failures - The Collection
Constructor: A fix to MakeWith_AllMedia to address the assert 
failures in Christmas pack '08. When going from a video to 
a picture and the duration of the picture is not long enough
to accommodate the ending transition set for the video, the
transition has to be shortened. There was a formula error
in computing this adjustment.