0014043: Analysis of 1 second silent wave file causes the application to crash. Eliminated the following assert failure, which was observed while investigating this issue: MDP2.cpp(444) ASSERT FAILED: (NULL == pAudioInfo)
The assert failed only if C:\muveeDebug\DumpDescriptors was present. It was due to call to IMusicData::GetAudioInfo in StyleInterfaceUtils::DumpMusicDescriptor with an IAudioInfo pointer not initialized to NULL (as expected by IMusicData::AudioInfo).

Turns out IAudioInfo was used only in DumpMusicDescriptor to print ramp up/down times, which aren't used during construction. Since the values aren't used, I've removed them from the descriptor dump entirely. With this change, the IAudioInfo interface is no longer used anywhere in the constructor code.