#0020710: Output shows jagged edges when production is saved to AVC-HD at 1920x1080 resolution
16x multisampling of AVC-HD saving causes reveal to crash on my PC. I have an NVIDIA Quadro FX 540. I tried saving a muvee to AVC-HD at 1920x1080 using cube twist with 4x multisampling and the saved muvee is generally ok. There are indeed jaggy edges but I found them to be acceptable.

That being said, some people with super high-end cards might insist on 16x. So 16x multisampling has been added when saving to any format. It is not enabled by default. If the file MultiSampling16x.txt is present in c:\muveeDebug, then we'll try 16x and progressively decrease the sampling rate in case of failure.