#0022867: Need for Cg Compiler
The Cg Compiler has been moved to CgFragmentShader.cpp/.h. So now all future effects will use this generic compiler.
First drop of CgFragmentProgram effect. It is a generic Cg effect has allows the program to be sent as a ProgramString from muse. So far it only has 3 float parameters. It will add a whole variety of parameters soon.
The ConvertToASCII( const FString *program ) has been inserted in EffectUtils.cpp It is widely used and will be extensively referenced in the future.
Deleted CgTest effect. We no longer need it as it was a proof of concept for our Cg compiler.