#0022792: S00538_LifeStory
Merging with the trunk rev8127. 

Please note that this is a MANUAL merge of BlurLayers effect and a helper function in EffectUtils. I have excluded the Cg compiler from this merge. There are two main reasons for that. The first being that BR2 does not need the Cg compiler. The second being that The Cg compiler needs to be merged all over the place in BR2. I think its wiser to wait until the BR3 is created and is split from the trunk ver8161 or beyond.