Resolved: #22925: Updated the algo used in font dialog's listbox intelligent mapping to the selection textbox entry. It's to facilitate user to locate the list entry via typing in the textbox. Here we did not distinguish if entry is user input or due to listbox selection, hence the mapping is two-way.
Previously the algo will find the first string that contains the text-box entry, hence SimSun will get directed to NSimSun in the listbox and this will cause upwards shift in the listbox, when user continues to scroll down with the down key, he will reach SimSun again and the the scroll back up to NSimSun, causing the erratic scrolling behaviour. Fix here is to make sure the match starts with the text-box entry instead.