0021182: Video Summarized, 1 Clip Dropped. The First Clip was Dropped not the Last.
The constructor (prior to this commit) will try to show at least
one segment from each file given in the input. With this commit,
it will try to show at least one segment for each thumbnail 
that the user sees in the media panel. This is clearer from the
user's perspective as well as gives us an extra feature - use
at least one segment from every shot. All you have to do to
get one segment for every shot is to split your video into
shots in the media panel.