0023024: extremely jerky muvee preview (with video sources)
The VideoFileWithDelay approach was introduced for better handling
of multiple input video files. However, it also introduces
an extra VideoIO thread which causes extra load 
in terms of thread switching on single core CPUs.

Right now in trunk, we have a "DisableVideoDelay.txt" flag 
which can be used to turn on and off the VideoFileWithDelay approach. 
Kumar and I think that we should go for the old VideoFileImpl approach 
on single cores which gives smooth preview. 

The down side of this is that single core computers 
will run out of memory easily (with 10 or 20++ video inputs).