0022948: [SpringRefresh09] FSP is very jerky on HP TouchSmart
This checkin addresses the problem mentioned in my (0064479) by making not to use IDirectDraw->->SetCooperativeLevel() as default. It also uses GPUCapabilities.txt instead of "c:\\muveeDebug\\FS_NoIDirectDraw.txt" to signal this setting on and off.

Usage: 
Add line {("MVLOCALHOST" use-idirectdraw-for-fullscreen)} or {("Some specific render" use-idirectdraw-for-fullscreen)} in GPUCapabilities.txt to enable the use of IDirectDraw for full screen preview. This can be used to profile the renderers in future.