0022917: [Mini] Savng to MPEG-2 format will cause application to crash 0022915: [Mini] Savng to MPEG-1 format will cause application to crash 0022894: [Support FPU-25977-292] Crash on burning DVD
There is need for synchornization between
SaveRenderer and DirectShow SavingGraph to run
so that DirectShow saving shouldn't start
untill Save-OPENGL context is created. Previously
we set the time out limit set to 2 seconds but this
timeout limit still get expired on some machine so we 
increased this timeout limit to 10 seconds.