#0017740: Need Drop Shadow primitive
In order to allow a shadow to be projected on an arbitrary surface, a new effect is needed. The new effect, Stencil, simply interacts with the openGL stencil buffer before we draw our custom surface. Then when we reach the Shadow effect further down in the stack, we have an active buffer to use. The stencil effect has been added to ShowTime_FX because of its simplicity.

stylelib2 has been modified to contain some stencil buffer constants.
Shadow effect has an additional parameter called UseBackground with is set to 1 by default. That means we'll use our internal surface to project the shadow on unless the style author decides to use a custom surface.
Shadow effect code has also been reorganized.